add mysql como banco de dados e adicionando validação de usuário

+++ app.py
@@ -11,7 +11,9 @@ from flask_login import (
 
 app = Flask(__name__)
 app.config["SECRET_KEY"] = "your_secret_key"
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///database.db"
+app.config["SQLALCHEMY_DATABASE_URI"] = (
+    "mysql+pymysql://root:admin123@127.0.0.1:3306/flask-crud"
+)
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -53,7 +55,7 @@ def create_user():
     password = data.get("password")
 
     if username and password:
-        user = User(username=username, password=password)
+        user = User(username=username, password=password, role="user")
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "Usuário cadastrado com sucesso!"})
@@ -77,6 +79,9 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"message": "Operação não permitida"}), 403
+
     if user and data.get("password"):
         user.password = data.get("password")
         db.session.commit()
@@ -91,8 +96,11 @@ def update_user(id_user):
 def delete_user(id_user):
     user = User.query.get(id_user)
 
+    if current_user.role != "admin":
+        return jsonify({"message": "Operação não permitida."}), 403
+
     if id_user == current_user.id:
-        return jsonify({"message": "Deleção não permitida"}), 404
+        return jsonify({"message": "Deleção não permitida"}), 403
 
     if user:
         db.session.delete(user)
